Combine results of queries based on partitions in class property
