+++ kuwala-pipelines/google-poi/src/pipeline/main.py
@@ -1,7 +1,8 @@
-from search_scraper import scrape_with_search_string
+from SearchScraper import SearchScraper
 from search_string_generator import generate_search_strings
 
 if __name__ == '__main__':
     search_strings = generate_search_strings()
+    search_scraper = SearchScraper()
 
-    scrape_with_search_string(search_strings)
+    search_scraper.scrape_with_search_string(search_strings)
